ANDR-687
The task is done.

+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/objectstore/StoredObject.java
@@ -9,6 +9,7 @@ import android.os.Parcel;
 import android.os.Parcelable;
 
 import com.fieldnation.GlobalState;
+import com.fieldnation.Log;
 import com.fieldnation.service.objectstore.ObjectStoreSqlHelper.Column;
 import com.fieldnation.utils.misc;
 
@@ -628,7 +629,8 @@ public class StoredObject implements Parcelable, ObjectStoreConstants {
     }
 
     public static boolean delete(Context context, StoredObject obj) {
-        if (obj.isFile()) {
+
+        if (obj != null && obj.isFile()) {
             obj.getFile().delete();
         }
